refactor(viewer): split formatLoadError out of load-errors.ts
`load-errors.ts` sat at 401 lines, over the ~400-line module rule, so the
anchoring fix that follows had nowhere to land. The two halves have
different audiences and different blast radii — the classifier feeds
analytics grouping and the severity/drop decisions in `analytics-scrub.ts`,
`formatLoadError` only ever produces a toast string — so this moves the
humanisation half to a sibling `load-error-message.ts` and leaves the next
commit's diff about matching semantics alone.

Pure move: no re-export shim, the three importers updated in place, and
`messageOf` / `WASM_RUNTIME_UNRECOVERABLE_MARKER` exported rather than
duplicated so the marker keeps one definition.
