+++ apps/viewer/src/components/viewer/GLBExportDialog.tsx
@@ -44,7 +44,8 @@ import { buildHiddenIfcTypes } from '@/store/typeVisibilityFilter';
 import { posthog } from '@/lib/analytics';
 import { toast } from '@/components/ui/toast';
 import { GeometryProcessor, isNoRenderGeometryError, type MeshData } from '@ifc-lite/geometry';
-import { classifyLoadError, formatLoadError } from '@/lib/load-errors';
+import { classifyLoadError } from '@/lib/load-errors';
+import { formatLoadError } from '@/lib/load-error-message';
 import { exportGlbFromGeometry } from '@/lib/export/glb';
 import { downloadBlob, sanitizeFilename } from '@/lib/export/download';
 import { withInstancedMeshes } from '../../utils/instancedExport.js';
