+++ apps/viewer/src/lib/load-errors.ts
@@ -3,7 +3,8 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 /**
- * Classification + humanisation of model-load failures.
+ * Classification of model-load failures. The user-facing humanisation lives in
+ * ./load-error-message.ts.
  *
  * Despite the name, this is the viewer's ONLY error-family classifier —
  * `analytics-scrub.ts` runs it over every captured `$exception` — so a non-load
@@ -16,8 +17,9 @@
  * is not ok`. That message is meaningless to a user and, captured raw, is
  * hard to triage in error tracking.
  *
- * This module maps such failures to a stable `kind` (for analytics
- * grouping) and a human-readable message (for the toast / model loadError).
+ * This module maps such failures to a stable `kind` for analytics grouping;
+ * `formatLoadError` in ./load-error-message.ts turns that kind into the
+ * human-readable message (the toast / model loadError).
  */
 
 // Imported, not restated: a second copy of MapLibre's wordings would drift on
@@ -78,7 +80,7 @@ export type LoadErrorKind =
   | 'network_unavailable'
   /**
    * The browser refused a WebGL context to the location minimap (#2354). Not a
-   * load failure and never reaches {@link formatLoadError}; classified so the
+   * load failure and never reaches `formatLoadError`; classified so the
    * family gets ONE fingerprint instead of one issue per deploy. Membership is
    * by MESSAGE, not by who caught it: `isWebglContextCreationError` matches only
    * MapLibre's own wordings and the two strings `LocationMap` synthesizes, all
@@ -101,7 +103,8 @@ function errorNameOf(err: unknown): string {
   return typeof name === 'string' ? name : '';
 }
 
-function messageOf(err: unknown): string {
+/** Exported for ./load-error-message.ts, which needs the same stringification. */
+export function messageOf(err: unknown): string {
   if (err instanceof Error) return err.message;
   if (typeof err === 'string') return err;
   return String(err);
@@ -212,7 +215,7 @@ function isGeometryWorkerCrashError(message: string): boolean {
  * object carries its `.name`, so nothing real is lost. Checked AFTER the worker
  * bucket so a worker-attributed trap keeps its own bucket.
  */
-const WASM_RUNTIME_UNRECOVERABLE_MARKER = 'WASM_RUNTIME_UNRECOVERABLE'; // == @ifc-lite/geometry's WASM_RUNTIME_UNRECOVERABLE_CODE
+export const WASM_RUNTIME_UNRECOVERABLE_MARKER = 'WASM_RUNTIME_UNRECOVERABLE'; // == @ifc-lite/geometry's WASM_RUNTIME_UNRECOVERABLE_CODE
 function isWasmRuntimeCrashError(err: unknown, message: string): boolean {
   return (
     errorNameOf(err) === 'RuntimeError' || message.includes(WASM_RUNTIME_UNRECOVERABLE_MARKER)
@@ -328,74 +331,4 @@ export function errorCaptureProps(err: unknown): Record<string, unknown> {
   const nav = (globalThis as { navigator?: { onLine?: unknown } }).navigator;
   if (typeof nav?.onLine === 'boolean') props.online = nav.onLine;
   return props;
-}
-
-/**
- * Produce a user-facing message for a load failure. Known failure modes get
- * actionable guidance; everything else falls back to the raw error text so we
- * never hide useful detail.
- *
- * @param fileName Optional file name to attribute the failure to.
- */
-export function formatLoadError(err: unknown, fileName?: string): string {
-  const kind = classifyLoadError(err);
-  const subject = fileName ? `"${fileName}"` : 'the model';
-  switch (kind) {
-    case 'wasm_engine_load':
-      return (
-        `Couldn't load the 3D geometry engine — a required file failed to download. ` +
-        `This usually means the app updated in the background, or a proxy/antivirus blocked it. ` +
-        `Please reload the page (Ctrl/Cmd+Shift+R). If it persists, check your network or extensions.`
-      );
-    case 'out_of_memory':
-      return (
-        `Ran out of memory while processing ${subject}. ` +
-        `Try closing other tabs, or load fewer/smaller models at once.`
-      );
-    case 'geometry_worker_crash':
-      return (
-        `A geometry worker stopped unexpectedly while processing ${subject}. ` +
-        `This usually means the model is too large for this device's available memory. ` +
-        `Try closing other tabs, or load fewer/smaller models at once.`
-      );
-    case 'geometry_stream_stalled':
-      return (
-        `Processing ${subject} stalled and was stopped. ` +
-        `The model may be too large or complex for this device. ` +
-        `Try closing other tabs, or load fewer/smaller models at once.`
-      );
-    case 'file_unreadable':
-      return (
-        `Couldn't read ${subject} — the file is no longer available to the browser. ` +
-        `It may have been moved, renamed, deleted, or unloaded by a cloud-sync client ` +
-        `(OneDrive/Dropbox/iCloud) since you picked it. Please select the file again.`
-      );
-    case 'wasm_runtime_crashed':
-      // Two sub-cases, and the difference matters to the user: a trap taken by
-      // one operation costs only that operation (the engine rebuilds itself on
-      // the next one), while a trap taken while the engine was starting cannot
-      // be undone without a new document. Never show the raw engine text here —
-      // the reported occurrence put an internal sentence about "recreating the
-      // worker process" in front of a user (#1898).
-      return messageOf(err).includes(WASM_RUNTIME_UNRECOVERABLE_MARKER)
-        ? (
-          `The 3D geometry engine crashed and can't restart in this tab. ` +
-          `Please reload the page (Ctrl/Cmd+R) — your work in other tabs is unaffected. ` +
-          `If it happens again on the same model, it is likely too large for this device's memory.`
-        )
-        : (
-          `The 3D geometry engine crashed while processing ${subject} and the operation was stopped. ` +
-          `This is usually memory pressure on a large or complex model — try closing other tabs, ` +
-          `or exporting/loading a smaller selection. Reload the page if it keeps happening.`
-        );
-    case 'cancelled':
-      return `Loading ${subject} was cancelled.`;
-    case 'network_unavailable':
-      return (
-        `Couldn't load ${subject}: the connection dropped while downloading. ` +
-        `Check your network and try again. Nothing was lost, so loading the same file again is safe.`
-      );
-    default:
-      return `Failed to load ${subject}: ${messageOf(err)}`;
-  }
 }
